Moved rustup related functions to rustup.ts
